### 2019-03-10 #1311 @tknopp
Hi, I am very excited about JOSS and this is my first submission. Getting a domain expert for review will be somewhat challenging. I have found one reviewer with expertise in magnetic resonance imaging, which would be close enough. But maybe its more important that there is Julia expertise. In the end, my package is basically a file loading / storing library.
