### 2019-03-16 #1311 @brainstorm
Sure! While I'm here I would like to ask for guidance on my QKD quantum
paper... The reviewer seems to have ghosted us mid-review, what's the best
course of action here?

Den sön 17 mars 2019 08:22Lorena A. Barba <notifications@github.com> skrev:

> @brainstorm <https://github.com/brainstorm> — Taking into account editor
> workloads at the moment, I'd like to assign this submission to you for
> handling. Can you take it on?
>
