### 2019-03-18 #1331 @whedon
[REVIEW]: MPIFiles.jl: A Julia Package for Magnetic Particle Imaging Files
**Submitting author:** @tknopp (<a href="http://orcid.org/0000-0002-1589-8517">Tobias Knopp</a>)
**Repository:** <a href="https://github.com/MagneticParticleImaging/MPIFiles.jl" target ="_blank">https://github.com/MagneticParticleImaging/MPIFiles.jl</a>
**Version:** v0.5.3
