### 2019-03-27 40b8c76
fix bug with averaging of nonmatching block length
