### 2019-04-02 7d3c13a
fix different numPeriodPerFrame for Multipatch reco
