### 2019-04-06 79711bf
implement BrukerFile reco data loading
