+++ src/RecoData.jl
@@ -1,13 +1,13 @@
 # MDF reconstruction data is loaded/stored using ImageMeta objects from
 # the ImageMetadata.jl package.
 
-export converttometer, imcenter, loadRecoDataMDF, saveRecoDataMDF
+export converttometer, imcenter, loadRecoData, saveRecoData
 
 converttometer(x) = ustrip.(uconvert.(u"m",x))
 imcenter(img::AxisArray) = map(x->(0.5*(last(x)+first(x))), ImageAxes.filter_space_axes(AxisArrays.axes(img), axisvalues(img)))
 imcenter(img::ImageMeta) = imcenter(data(img))
 
-function saveRecoDataMDF(filename, image::ImageMeta)
+function saveRecoData(filename, image::ImageMeta)
   C = colordim(image) == 0 ? 1 : size(image,colordim(image))
   L = timedim(image) == 0 ? 1 : size(image,timedim(image))
 
@@ -34,12 +34,12 @@ function saveRecoDataMDF(filename, image::ImageMeta)
   end
 end
 
-function loadRecoDataMDF(filename::AbstractString)
-  f = MDFFile(filename)
-  return loadRecoDataMDF(f)
+function loadRecoData(filename::AbstractString)
+  f = MPIFile(filename)
+  return loadRecoData(f)
 end
 
-function loadRecoDataMDF(f::MDFFile)
+function loadRecoData(f::MDFFile)
   header = loadMetadata(f)
   header["datatype"] = "MPI"
 
@@ -47,13 +47,13 @@ function loadRecoDataMDF(f::MDFFile)
   if recoParams != nothing
     header["recoParams"] = recoParams
   end
-  im = loadRecoDataMDF_(f)
+  im = loadRecoData_(f)
   imMeta = ImageMeta(im, header)
 
   return imMeta
 end
 
-function loadRecoDataMDF_(f::MDFFile)
+function loadRecoData_(f::MDFFile)
   # preparation for spatial axes
   rsize::Vector{Int64} = recoSize(f)
   pixspacing = (recoFov(f) ./ rsize)*1000u"mm"
@@ -74,13 +74,26 @@ function loadRecoDataMDF_(f::MDFFile)
   # load data
   c_::Array{Float32,3} = recoData(f)
   c::Array{Float32,5} = reshape(c_, size(c_,1), rsize[1], rsize[2], rsize[3], size(c_,3))
-  
-  return AxisArray(c, Axis{:color}(1:size(c,1)),
-		   Axis{:x}(range(offset[1], step=pixspacing[1], length=size(c,2))),
-		   Axis{:y}(range(offset[2], step=pixspacing[2], length=size(c,3))),
-		   Axis{:z}(range(offset[3], step=pixspacing[3], length=size(c,4))),
-		   Axis{:time}(range(0.0u"s", step=periodTime, length=size(c,5))))
+
+  return makeAxisArray(c, pixspacing, offset, periodTime)
 end
 
-#precompile(MPIFiles.loadRecoDataMDF_,(MDFFileV1,))
-#precompile(MPIFiles.loadRecoDataMDF_,(MDFFileV2,))
+
+function loadRecoData(f::BrukerFile)
+  # preparation for spatial axes
+  rsize::Vector{Int64} = recoSize(f)
+  pixspacing = (recoFov(f) ./ rsize)*1000u"mm"
+  off::Vector{Float64} = vec(recoFovCenter(f))
+  offset = [0.0,0.0,0.0]*u"mm"
+  if off != nothing
+    offset[:] = (off .- 0.5.*recoFov(f))*u"m" .+ 0.5.*pixspacing
+  end
+
+  periodTime = 1.0*u"s"
+
[…]
