+++ src/Utils.jl
@@ -1,3 +1,5 @@
+export makeAxisArray
+
 
 # Support for handling complex datatypes in HDF5 files
 function writeComplexArray(file, dataset, A::AbstractArray{Complex{T},D}) where {T,D}
@@ -66,3 +68,13 @@ function str2uuid(str::String)
   end
 end
 str2uuid(str::Nothing) = str
+
+function makeAxisArray(array::Array{T,5}, pixelspacing, offset, dt) where T
+  N = size(array)
+  im = AxisArray(array, Axis{:color}(1:N[1]),
+		 Axis{:x}(range(offset[1],step=pixelspacing[1],length=N[2])),
+		 Axis{:y}(range(offset[2],step=pixelspacing[2],length=N[3])),
+		 Axis{:z}(range(offset[3],step=pixelspacing[3],length=N[4])),
+		 Axis{:time}(range(0*unit(dt),step=dt,length=N[5])))
+  return im
+end
