### 2019-05-13 cbf0a97
small fixes
