### 2019-05-15 c331f90
fix timeString for open measurments
