### 2019-05-16 #1331 @tknopp
@labarba: If there is anything I can do, please let me know. I have a depending paper under review and would of course like to reference the software being used. I could ask in the Julia forum for potential reviewers if @IgorBaratta has no time. Or maybe you have other reviewers in mind. Thanks!
