### 2019-05-18 #1331 @IgorBaratta
Hi @tknopp, @labarba,  sorry for not addressing this review sooner!

I found the package quite practical, primarily because of the possibility of reading different file formats in a common interface. I've tested the package functionality with MDF data-sets available in the OpenMPIData repository and the Bruker MPI data-sets used in the tests.

Everything worked out smoothly.  Though it took some hours for downloading and testing the big data-sets, as expected.

However, I'd suggest some improvements before recommending acceptance.

**Software Paper:**

- The pdf file has lost its formatting. Please check the 3rd paragraph.
- I would also recommend checking the structure of the 2nd paragraph to improve readability.

**Documentation:**

- Example usage -  I didn't find any clear example in the documentation. I used the data-sets mentioned above to verify the core functionality of the package, but they are not documented. I'd suggest include at least one example with an actual (small) data-set, maybe in the Getting Started section.

- Community guidelines – Although there is a clear indication of the issue tracker (to report issues and seek support), I was not able to find any guidelines for contributing to the package.

I hope this review would help improve the package.
