### 2019-05-19 #1331 @tknopp
Thank you @IgorBaratta! These are all very good points that definitely need to be addressed. I will work on this and ping you when its ready.
