+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @Book{knopp2012magnetic,
   Title                    = {Magnetic Particle Imaging: An Introduction to Imaging Principles and Scanner Instrumentation},
-  Author                   = {Tobias Knopp and {Thorsten M. Buzug}},
+  Author                   = {Tobias Knopp and {T. M.} Buzug},
   Publisher                = {Springer},
   Year                     = {2012},
   Address                  = {Berlin/Heidelberg},
