+++ paper/paper.md
@@ -31,13 +31,13 @@ affiliations:
    index: 1
  - name: Institute for Biomedical Imaging, Hamburg University of Technology
    index: 2
-date: ?? March 2019
+date: ?? May 2019
 bibliography: paper.bib
 ---
 
 # Summary
 
-Tomographic imaging methods allows to visualize the interior of the human body in a
+Tomographic imaging methods allow to visualize the interior of the human body in a
 non-invasive way. Most prominent examples used for medical diagnosis are the
 computed tomography (CT) and magnetic resonance imaging (MRI).
 In 2015 [@Gleich2005], Bernhard Gleich and Jürgen Weizenecker introduced a new tomographic
@@ -47,19 +47,9 @@ The MNP can be injected intravenously and allow to image the vessel tree as well
 as organ perfusion. The MNPs are harmless for the human body and are degraded in the
 liver.
 
-When working with an MPI scanner there are various datasets involved. During an
-MPI experiment, the signal is measured using inductive receive coils. In order to
-reconstruct an image of the MNP distribution one requires knowledge of the MPI
-system matrix, which is usually measured within a calibration measurement.
-The software package ``MPIFiles.jl`` provides an interface to all these types of data.
-In particular it has full support for the Magnetic Particle Imaging
-Data Format (MDF) [@knopp2016MDF], which is an open, HDF5 based, data format that
-has been developed with the purpose of standardizing data storage in MPI. Additionally,
-``MPIFiles.jl`` has full read support for datasets measured with the preclinical
-MPI scanner from the vendor Bruker. Each data format is implemented using a dedicated
-type and implements a common interface such that the user of the package can write
-generic code for different file types. In addition to read support, the package also
-has conversion routines for creating MDF files from Brukerfiles.
+When working with an MPI scanner different datasets are involved. In addition to the raw data, which consist of the induced voltage signals, a calibration dataset is required to reconstruct the particle concentration. The latter describes the physical relationship between the particle concentration and the measurement signal and is usually recorded in MPI with a delta-shaped point sample.
+The software package ``MPIFiles.jl`` provides an interface to all these data types. Besides the proprietary data format recorded with MPI scanners from Bruker, ``MPIFiles.jl`` implements the vendor-independent Magnetic Particle Imaging
+Data Format (MDF)[@knopp2016MDF], which is based on HDF5 and was developed with the aim to standardize data storage in MPI. Each data format implements a common interface and enables the user to write generic code for different file types. Besides the read and write capability for different data formats, ``MPIFiles.jl`` also offers the possibility of converting different data formats into each other.
 
 ``MPIFiles.jl`` supports the following types of MPI data
 
@@ -68,14 +58,12 @@ has conversion routines for creating MDF files from Brukerfiles.
 * reconstruction results
 
 In addition to a low level access to the data, the package also features
-high level routines that provide various post-processing methods such as
+high level routines providing various post-processing methods such as
 frequency filtering, spectral leakage correction, block-averaging, and
 transfer function correction.
 
-The API of ``MPIFiles.jl`` is aligned with the parameter names of the MDF
-specification such that the latter can be used as a reference for determining
-the content and dimensionality of certain parameters.
-``MPIFiles.jl`` is licensed under MIT license and is hosted on Github at
+The API of ``MPIFiles.jl`` is based on the parameter names of the MDF specification, so that the latter can be used as a reference for the description of parameters and their dimensionality.
+``MPIFiles.jl`` is licensed under the MIT license and hosted on Github at
 https://github.com/MagneticParticleImaging/MPIFiles.jl.
 
 # References
