+++ paper/paper.bib
@@ -43,6 +43,7 @@
     title = "{MDF: Magnetic Particle Imaging Data Format}",
   journal = {ArXiv e-prints},
 archivePrefix = "arXiv",
+   volume = {1602.06072}
    eprint = {1602.06072},
  primaryClass = "physics.med-ph",
  keywords = {Physics - Medical Physics},
