+++ paper/paper.bib
@@ -24,6 +24,16 @@
   timestamp = {2007.12.12}
 }
 
+@article{graser2019human,
+  title={Human-sized magnetic particle imaging for brain applications},
+  author={Gr{\"a}ser, Matthias and Thieben, Florian and Szwargulski, Patryk and Werner, Franziska and Gdaniec, Nadine and Boberg, Marija and Griese, Florian and Hofmann, Martin and Ludewig, Peter and Ven, Dionys van de and Weber, {O.M.} and Woywode, Oliver and Gleich, Bernhard and Knopp Tobias},
+  journal={Nature communications},
+  volume={10},
+  number={1936}
+  pages={1--9},
+  year={2019},
+  doi={10.1038/s41467-019-09704-x}
+}
 
 @article{knopp2017magnetic,
   title={Magnetic particle imaging: From proof of principle to preclinical applications},
