### 2019-05-23 090303e
improve docu
