### 2019-05-23 #1331 @tknopp
@IgorBaratta: I have worked on all of your points. Regarding this
> Example usage - I didn't find any clear example in the documentation. I used the data-sets mentioned above to verify the core functionality of the package, but they are not documented. I'd suggest include at least one example with an actual (small) data-set, maybe in the Getting Started section.

I was not 100% sure how this was meant. The dataset used in "getting started" was actually downloaded when performing running the test suite but I have not mentioned it in the documentation. Now I explicitly download the data:
    https://magneticparticleimaging.github.io/MPIFiles.jl/dev/gettingStarted.html
Then the user can directly try out all the exampled on that page.
