### 2019-05-28 9722ce0
Changes Access to frequency index
