### 2019-05-28 ebda40a
additional fix for datatype
