### 2019-06-07 71c487d
fix pseudo 2D freq selection
