### 2019-06-08 #1331 @IgorBaratta
The author has addressed all points.
To me this paper is ready for acceptance.

> Hi all — are we waiting here for the reviewer to complete their tasks? or the author to do some revisions? Or both? Give us an update!
