### 2019-06-09 #1331 @brainstorm
I see a couple of points unchecked in the checklist: example usage and community guidelines?
