### 2019-06-09 #1331 @IgorBaratta
> I see a couple of points unchecked in the checklist: example usage and community guidelines?
My mistake. I guess now they are all checked.
