### 2019-06-09 #1331 @tknopp
@brainstorm: I am not sure what you mean by "register a zenodo DOI", could you elaborate? Should I register a DOI for one of the references or for the article itself? Would be great to get some help.
