### 2019-06-09 #1331 @labarba
I've read the paper and made some editorial suggestions via this PR
https://github.com/MagneticParticleImaging/MPIFiles.jl/pull/14
