### 2019-06-09 #1331 @labarba
Note: after making the Zenodo deposit, be sure to check that it has the correct metadata, and edit as necessary. This includes the title (should match the paper title) and author list (make sure the list is correct and people like me who only fixed some bits of the paper, are not on it). You may also want to add the ORCID of all authors in the Zenodo metadata.
