### 2019-06-09 #1331 @tknopp
Thanks @labarba, I have merged your changes and initiated a release of the package:

https://github.com/JuliaRegistries/General/pull/1295

Note that this is a two-step procedure in the Julia universe and the github release is done in the second step. I thus need to wait until the upper PR is merged. Then I will create the zenodo DOI.
