### 2019-06-10 #1331 @tknopp
@labarba: I hope I got this right:
- Here is the Zenodo release: https://zenodo.org/record/3242352#.XP4-Uy2BrLE
- The DOI is: 10.5281/zenodo.3242352
- MPIFiles has now Version 0.5.3
