### 2019-06-10 #1331 @tknopp
Awesome, thank you very much @labarba, @brainstorm and @IgorBaratta. Will celebrate my first JOSS paper now!
